+++ src/LSST/M1M3/TS/Commands/EnterControl.h
@@ -43,7 +43,7 @@ public:
         TSPublisher::instance().logSoftwareVersions();
         TSPublisher::instance().logSimulationMode();
 
-        Events::SummaryState::setState(MTM1M3TS::MTM1M3TS_shared_SummaryStates_StandbyState);
+        Events::SummaryState::set_state(MTM1M3TS::MTM1M3TS_shared_SummaryStates_StandbyState);
 
         return Task::DONT_RESCHEDULE;
     }