+++ src/LSST/M1M3/TS/Events/SummaryState.cpp
@@ -44,23 +44,23 @@ void SummaryState::send() {
 }
 
 bool SummaryState::active() {
-    std::lock_guard<std::mutex> lockG(_stateMutex);
+    std::lock_guard<std::mutex> lockG(_state_mutex);
     return summaryState == MTM1M3TS_shared_SummaryStates_DisabledState ||
            summaryState == MTM1M3TS_shared_SummaryStates_EnabledState;
 }
 
 bool SummaryState::enabled() {
-    std::lock_guard<std::mutex> lockG(_stateMutex);
+    std::lock_guard<std::mutex> lockG(_state_mutex);
     return summaryState == MTM1M3TS_shared_SummaryStates_EnabledState;
 }
 
-void SummaryState::_switchState(int newState) {
-    std::lock_guard<std::mutex> lockG(_stateMutex);
-    SPDLOG_TRACE("SummaryState::_switchState from {} to {}", summaryState, newState);
-    if (summaryState == newState) {
-        throw std::runtime_error(fmt::format("Already in {} state", newState));
+void SummaryState::_switch_state(int new_state) {
+    std::lock_guard<std::mutex> lockG(_state_mutex);
+    SPDLOG_TRACE("SummaryState::_switch_state from {} to {}", summaryState, new_state);
+    if (summaryState == new_state) {
+        throw std::runtime_error(fmt::format("Already in {} state", new_state));
     }
-    switch (newState) {
+    switch (new_state) {
         case MTM1M3TS_shared_SummaryStates_DisabledState:
             if (summaryState != MTM1M3TS_shared_SummaryStates_StandbyState &&
                 summaryState != MTM1M3TS_shared_SummaryStates_EnabledState) {
@@ -77,8 +77,11 @@ void SummaryState::_switchState(int newState) {
             break;
         case MTM1M3TS_shared_SummaryStates_OfflineState:
             if (summaryState != MTM1M3TS_shared_SummaryStates_StandbyState) {
-                throw std::runtime_error(fmt::format("Cannot switch to Offline state from {}", summaryState));
+                SPDLOG_ERROR("Ignored request to switch to offline state from non-Standby state - {}.",
+                             summaryState);
+                return;
             }
+            break;
         case MTM1M3TS_shared_SummaryStates_StandbyState:
             if (summaryState != MTM1M3TS_shared_SummaryStates_DisabledState &&
                 summaryState != MTM1M3TS_shared_SummaryStates_OfflineState) {
@@ -86,9 +89,9 @@ void SummaryState::_switchState(int newState) {
             }
             break;
         default:
-            throw std::runtime_error(fmt::format("Invalid target state: {}", newState));
+            throw std::runtime_error(fmt::format("Invalid target state: {}", new_state));
     }
     instance()._updated = true;
-    summaryState = newState;
+    summaryState = new_state;
     send();
 }