+++ src/LSST/M1M3/TS/Telemetry/ThermalData.h
@@ -47,6 +47,8 @@ public:
      * Sends updates through SAL/DDS.
      */
     void send();
+
+    auto get_absoluteTemperature() { return absoluteTemperature; }
 };
 
 }  // namespace Telemetry